Added prototype for Square which inherits from Rectangle.  Added file to create prototype for an Elliptical body.  Added conditionals to the collision detection methed for different body shapes.
Added collision detection and setting of normal vector for limited case of
collision of two squares of normal orientaion.  Added a variable to get the
orientation of bodies of rectangle type for more general collisions.

Added the start of simple method for adding torque for off center
collisions of squares.  Work in progress.